requirements.txt
google-genai==0.4.0
requests==2.32.3
python-dotenv==1.0.1
rich==13.8.1
pyfiglet==0.8.1
Pillow==10.2.0
